// src/generated/accounts/loan.ts
export type Loan = {
  discriminator: ReadonlyUint8Array;
  version: number;
  /** amount in tokens, decimals included */
  principalLamports: bigint;
  orderBook: Address;
  /** Token mint for what the loan is in (spl address) */
  valueTokenMint: Address;
  escrowBumpSeed: number;
  /** stores start and duration */
  loanState: LoanState;
};